+++ src/main/java/kr/hhplus/be/server/interfaces/balance/api/BalanceController.java
@@ -1,8 +1,9 @@
-package kr.hhplus.be.server.interfaces.balance;
+package kr.hhplus.be.server.interfaces.balance.api;
 
-import kr.hhplus.be.server.application.balance.BalanceFacade;
 import kr.hhplus.be.server.application.balance.BalanceResult;
 import kr.hhplus.be.server.interfaces.ApiResponse;
+import kr.hhplus.be.server.interfaces.balance.api.BalanceResponse;
+import kr.hhplus.be.server.interfaces.balance.api.BalanceService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
 
@@ -12,18 +13,18 @@ import jakarta.validation.Valid;
 @RequestMapping("/api/v1/users")
 @RequiredArgsConstructor
 public class BalanceController {
-    private final BalanceFacade balanceFacade;
+    private final BalanceService balanceService;
 
     @GetMapping("/{userId}/balance")
     public ApiResponse<BalanceResponse.Balance> getBalance(@PathVariable Long userId) {
-        BalanceResult.Balance balance = balanceFacade.getBalance(userId);
+        BalanceResult.Balance balance = balanceService.getBalance(userId);
         return ApiResponse.success(BalanceResponse.Balance.of(balance));
     }
 
     @PostMapping("/{userId}/balance/charge")
     public ApiResponse<Void> chargeBalance(@PathVariable Long userId, 
                                            @Valid @RequestBody BalanceRequest.Charge request) {
-        balanceFacade.chargeBalance(request.toCriteria(userId));
+        balanceService.chargeBalance(request.toCriteria(userId));
         return ApiResponse.success();
     }
 }
