+++ src/test/java/kr/hhplus/be/server/support/ControllerTestSupport.java
@@ -8,17 +8,15 @@ import org.springframework.test.web.servlet.MockMvc;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
 import kr.hhplus.be.server.application.balance.BalanceFacade;
-import kr.hhplus.be.server.interfaces.balance.BalanceController;
-import kr.hhplus.be.server.interfaces.orders.api.OrderController;
 import kr.hhplus.be.server.application.order.OrderFacade;
 import kr.hhplus.be.server.application.product.ProductFacade;
-import kr.hhplus.be.server.interfaces.products.ProductController;
-
-import kr.hhplus.be.server.application.user.UserCouponFacade;
-import kr.hhplus.be.server.interfaces.user.UserCouponController;
-
 import kr.hhplus.be.server.application.rank.RankFacade;
+import kr.hhplus.be.server.application.user.UserCouponFacade;
+import kr.hhplus.be.server.interfaces.balance.api.BalanceController;
+import kr.hhplus.be.server.interfaces.orders.api.OrderController;
+import kr.hhplus.be.server.interfaces.products.ProductController;
 import kr.hhplus.be.server.interfaces.rank.RankController;
+import kr.hhplus.be.server.interfaces.user.UserCouponController;
 
 @WebMvcTest(controllers = {
     BalanceController.class,
